Replace string Actions with object Actions
